data base added

+++ src/main/java/com/torresj/nursing_sas_ope_info/services/NursesService.java
@@ -9,6 +9,8 @@ import java.util.Set;
 
 public interface NursesService {
     List<NurseOpeResponseDto> getOpeNurses(String filter) throws IOException;
-    NurseOpeResponseDto getOpeNurse(int id) throws IOException;
+
+    NurseOpeResponseDto getOpeNurse(long id) throws IOException;
+
     Set<NurseBolsaDto> getBolsaNurses(String filter) throws IOException;
 }
